enable to store the record_time after a game_end

+++ backend/typinggame/models.py
@@ -45,6 +45,6 @@ class AppUser(AbstractBaseUser, PermissionsMixin):
 
 class UserRecord(models.Model):
     user_name = models.CharField(max_length=255)
-    elapsed_time = models.FloatField()
+    elapsed_time = models.DecimalField(max_digits=5, decimal_places=2)
     timestamp = models.DateTimeField(auto_now_add=True)
     word_length = models.PositiveIntegerField()
